Turn getTotalDistance into totalEmissions
